JAX_irer-v14/orchestrator/requirements.txt
redis==5.0.1
fastapi
rq
numpy==1.26.0
jax
psycopg2-binary==2.9.9
sqlalchemy==2.0.15
alembic==1.7.7
pytest==7.2.0
pytest-asyncio==0.18.3
pytest-cov==3.0.0
pytest-mock==3.7.0
pytest-xdist==2.5.0
pytest-cases==3.0.0
pytest-redis==0.9.0
pytest-postgresql==4.0.0
pytest-asyncio==0.18.3
pytest-cov==3.0.0
pytest-mock==3.7.0
pytest-xdist==2.5.0
pytest-cases==3.0.0
pytest-redis==0.9.0
pytest-postgresql==4.0.0